Add confirmation for membership deletion
Change dataset constants to Link types from server data

+++ src/javascripts/redux/middlewares/teams.es6.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import getOrgId from 'redux/selectors/getOrgId.es6';
 import { Notification } from '@contentful/forma-36-react-components';
 import createTeamService from 'app/OrganizationSettings/Teams/TeamService.es6';
@@ -6,11 +7,12 @@ import { TEAM_MEMBERSHIPS, TEAMS } from '../dataSets.es6';
 import addCurrentTeamToMembership from 'redux/utils/addCurrentTeamToMembership.es6';
 import getDatasets from 'redux/selectors/getDatasets.es6';
 import { USERS } from 'redux/dataSets.es6';
+import ModalLauncher from 'app/common/ModalLauncher.es6';
+import RemoveMembershipConfirmation from 'app/OrganizationSettings/Teams/TeamMemberships/RemoveMembershipConfirmation.es6';
 
 export default ({ dispatch, getState }) => next => async action => {
   switch (action.type) {
     case 'SUBMIT_NEW_TEAM': {
-      next(action);
       const service = createTeamService(await getOrgId(getState()));
       const newTeam = await service.create(action.payload.team);
       dispatch({ type: 'ADD_TO_DATASET', payload: { item: newTeam, dataset: TEAMS } });
@@ -18,7 +20,6 @@ export default ({ dispatch, getState }) => next => async action => {
       break;
     }
     case 'SUBMIT_NEW_TEAM_MEMBERSHIP': {
-      next(action);
       const state = getState();
       const service = createTeamService(await getOrgId(state));
       const teamId = getCurrentTeam(state);
@@ -34,17 +35,24 @@ export default ({ dispatch, getState }) => next => async action => {
       break;
     }
     case 'REMOVE_TEAM_MEMBERSHIP': {
-      // get data from store before reducer runs and updates it
       const state = getState();
       const datasets = getDatasets(state);
       const service = createTeamService(await getOrgId(state));
-      next(action);
 
       const { teamMembershipId } = action.payload;
       const membership = datasets[TEAM_MEMBERSHIPS][teamMembershipId];
       const user = datasets[USERS][membership.sys.user.sys.id];
       const teamId = getCurrentTeam(state);
       const team = datasets[TEAMS][teamId];
+
+      const confirmation = await ModalLauncher.open(({ isShown, onClose }) => (
+        <RemoveMembershipConfirmation isShown={isShown} onClose={onClose} user={user} team={team} />
+      ));
+
+      if (!confirmation) {
+        break;
+      }
+      dispatch({ type: 'REMOVE_TEAM_MEMBERSHIP_CONFIRMED', payload: { teamMembershipId } });
 
       try {
         await service.removeTeamMembership(teamId, teamMembershipId);
@@ -62,7 +70,6 @@ export default ({ dispatch, getState }) => next => async action => {
       }
       break;
     }
-    default:
-      next(action);
   }
+  next(action);
 };
